@@ -14,25 +14,28 @@
   "author": "Digitransit Team <digitransit@hsl.fi>",
   "license": "(AGPL-3.0 OR EUPL-1.2)",
   "dependencies": {
-    "@mjaakko/react-responsive-grid": "0.3.7",
-    "gatsby": "^2.0.33",
+    "gatsby": "^2.9.4",
     "gatsby-plugin-react-helmet": "^3.0.9",
-    "gatsby-plugin-sharp": "^2.0.0",
+    "gatsby-plugin-sharp": "^2.1.9",
+    "gatsby-plugin-styled-components": "^3.0.7",
     "gatsby-plugin-typography": "^2.2.1",
     "gatsby-remark-autolink-headers": "^2.0.12",
-    "gatsby-remark-copy-linked-files": "^2.0.8",
+    "gatsby-remark-copy-linked-files": "^2.0.13",
     "gatsby-remark-custom-blocks": "^2.0.7",
-    "gatsby-remark-images": "^2.0.0",
+    "gatsby-remark-images": "^3.0.16",
     "gatsby-remark-prismjs": "^3.1.0",
     "gatsby-source-filesystem": "^2.0.0",
-    "gatsby-transformer-remark": "^2.0.0",
+    "gatsby-transformer-remark": "^2.4.0",
     "prismjs": "^1.15.0",
-    "react": "^16.5.1",
-    "react-dom": "^16.5.1",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
     "react-helmet": "^5.2.0",
     "react-prism": "^4.3.2",
-    "react-style-normalizer": "1.2.8",
     "react-typography": "^0.16.13",
+    "styled-components": "^4.3.1",
     "typography": "^0.16.17"
+  },
+  "devDependencies": {
+    "babel-plugin-styled-components": "^1.10.1"
   }
 }
